fix (front) : convert Videocall.jsx to VideoCall.tsx
